Moved wasm bindings into a new file (out of lib.rs)
